fix: cookie-based i18n, remove [locale] routing to fix redirect loop
