--- ui/src/app/[locale]/layout.tsx
@@ -1,109 +1,0 @@
-import "../globals.css";
-
-import type { Metadata } from "next";
-import { Geist_Mono } from "next/font/google";
-import { Manrope, Space_Grotesk } from "next/font/google";
-import { NextIntlClientProvider } from "next-intl";
-import { getMessages } from "next-intl/server";
-import { Suspense } from "react";
-
-import ChatwootWidget from "@/components/ChatwootWidget";
-import AppLayout from "@/components/layout/AppLayout";
-import PostHogIdentify from "@/components/PostHogIdentify";
-import { SentryErrorBoundary } from "@/components/SentryErrorBoundary";
-import SpinLoader from "@/components/SpinLoader";
-import { ThemeProvider } from "@/components/ThemeProvider";
-import { Toaster } from "@/components/ui/sonner";
-import { AppConfigProvider } from "@/context/AppConfigContext";
-import { OnboardingProvider } from "@/context/OnboardingContext";
-import { OrgConfigProvider } from "@/context/OrgConfigContext";
-import { TelephonyConfigWarningsProvider } from "@/context/TelephonyConfigWarningsContext";
-import { AuthProvider } from "@/lib/auth";
-
-
-const bodyFont = Manrope({
-  subsets: ["latin"],
-  display: "swap",
-  variable: "--font-body",
-});
-
-const displayFont = Space_Grotesk({
-  subsets: ["latin"],
-  display: "swap",
-  variable: "--font-display",
-});
-
-const geistMono = Geist_Mono({
-  variable: "--font-geist-mono",
-  subsets: ["latin"],
-});
-
-export const metadata: Metadata = {
-  title: "Sativoice Enterprise",
-  description: "Piattaforma Voice AI Enterprise per il mercato italiano",
-};
-
-export default async function LocaleLayout({
-  children,
-  params,
-}: {
-  children: React.ReactNode;
-  params: Promise<{ locale: string }>;
-}) {
-  const { locale } = await params;
-  const messages = await getMessages();
-
-  return (
-    <html lang={locale} className="dark" suppressHydrationWarning>
-      <head>
-        {/* Inline script to prevent flash of light theme - runs before React hydrates.
-            Dark is the locked default: only an explicit stored 'light' opts out. */}
-        <script
-          dangerouslySetInnerHTML={{
-            __html: `
-              (function() {
-                try {
-                  var theme = localStorage.getItem('theme');
-                  if (theme === 'light') {
-                    document.documentElement.classList.remove('dark');
-                  } else {
-                    document.documentElement.classList.add('dark');
-                  }
-                } catch (e) {
-                  document.documentElement.classList.add('dark');
-                }
-              })();
-            `,
-          }}
-        />
-      </head>
-      <body
-        className={`${bodyFont.variable} ${displayFont.variable} ${geistMono.variable} antialiased`}>
-        <NextIntlClientProvider messages={messages}>
-          <ThemeProvider attribute="class" defaultTheme="dark" enableSystem={false} disableTransitionOnChange>
-            <SentryErrorBoundary>
-              <AuthProvider>
-                <AppConfigProvider>
-                  <Suspense fallback={<SpinLoader />}>
-                    <OrgConfigProvider>
-                      <TelephonyConfigWarningsProvider>
-                        <OnboardingProvider>
-                          <PostHogIdentify />
-                          <AppLayout>
-                            {children}
-                          </AppLayout>
-                          <Toaster />
-                          <ChatwootWidget />
-                        </OnboardingProvider>
-                      </TelephonyConfigWarningsProvider>
-                    </OrgConfigProvider>
-                  </Suspense>
-                </AppConfigProvider>
-              </AuthProvider>
-            </SentryErrorBoundary>
-          </ThemeProvider>
-        </NextIntlClientProvider>
-      </body>
-    </html>
-  );
-}
