+++ ui/src/i18n/request.ts
@@ -1,6 +1,6 @@
 import { getRequestConfig } from "next-intl/server";
 
-import { routing } from "./i18n/routing";
+import { routing } from "./routing";
 
 export default getRequestConfig(async ({ requestLocale }) => {
   let locale = await requestLocale;
