+++ ui/src/i18n/routing.ts
@@ -4,7 +4,7 @@ import { defineRouting } from "next-intl/routing";
 export const routing = defineRouting({
   locales: ["it", "en"],
   defaultLocale: "it",
-  localePrefix: "always",
+  localePrefix: "never",
 });
 
 export const { Link, redirect, usePathname, useRouter, getPathname } =
